+++ changelog.md
@@ -1,7 +1,15 @@
 # Changelog
 
+* Advancements for extracting enchantments from items and applying to enchanting table.
+* Curses disenchanting 
+
 ## Upcoming
 * Enchanting Feature 
+  * Enchanting tables now knows no enchantments by default, it must learn them
+  * Enchantments learned now also learn the level, making it the maximum for that enchantment, so if you give them an Efficiency 3 book, the max level for Efficiency will be 3.
+  * Enchanting table can now learn curses, they cost 0 and increase the max cost
+  * Empowered items now increase the max level of enchantments by one over the maximum of the enchanting table
+    * This means that you can increase the max level above the limit on the enchanting table by using commands
   * Data pack now replaces experience bottles with enchanted cleansed lapis
 * Fixed Durium hammer taking almost no enchantments
 * Fixed Serene Seasons being a mandatory mod